Adiciona o campo de quantidade do produto na compra

+++ src/main/java/com/betaseven/lojaonline/domain/model/Produto.java
@@ -16,7 +16,7 @@ public class Produto {
     private Long idProduto;
     private String nome;
     private Double preco;
-    private int quantidade;
+    private Long quantidade;
 
     public Produto(Long idProduto) {
         this.idProduto = idProduto;
